Add swipe animation, add switch on double tap, change internal behavior
There are 3 types of animations now. The first is the red box, which always moves to the touched point. The second is the blue box, which moves depending on the device motion.
The last one is the green box, which gets a speed depending on a swipe distance. Changed MultiTouchArea to MouseArea in order to support double clicks easily.
The MouseArea is also a TouchArea (see documentation for more information) and the multi touch is not needed for now. With the double click the visible box can be changed in order to
show behavior one by one.
Removed the tutorial parts from the qml, because at the moment it is only distracting.